+++ src/components/paymentPage/pdtInfo/PdtInfo.tsx
@@ -1,20 +1,5 @@
 import { useEffect } from "react";
 import {
-  editABrand,
-  editOPrice,
-  editPCategory,
-  editPCnt,
-  editPCost,
-  editPDiscount,
-  editPSale,
-  editPName,
-  editPNo,
-  editPOption,
-  editPPrice,
-  editPThumbnail,
-  editKNo,
-} from "../../../redux/reducer/paymentSlice";
-import {
   RootState,
   useAppDispatch,
   useAppSelector,
@@ -24,25 +9,6 @@ import * as t from "./pdtInfo.style";
 
 const PdtInfo = (props: DbDataType) => {
   console.log(props.data);
-  const { aBrand } = useAppSelector((state: RootState) => state.paymentReducer);
-
-  useEffect(() => {
-    dispatch(editKNo(Number(props.data[0].keywordNo)));
-    dispatch(editPNo(props.data[0].id));
-    dispatch(editPThumbnail(props.data[0].thumbnail));
-    dispatch(editPCategory(props.data[0].category));
-    dispatch(editABrand(props.data[0].brand));
-    dispatch(editPName(props.data[0].name));
-    dispatch(editPCost(props.data[0].price));
-    dispatch(editPDiscount(props.data[0].discount));
-    dispatch(editPSale(props.data[0].discount === 0 ? false : true));
-    dispatch(editPOption(props.data[0].option));
-    dispatch(editPPrice(props.data[0].totalPrice));
-    dispatch(editPCnt(props.data[0].totalCnt));
-    dispatch(editOPrice(props.data[0].totalPrice));
-  }, []);
-
-  console.log(aBrand);
 
   const dispatch = useAppDispatch();
   return (
